Add files via upload
"This repository showcases my journey in learning the fundamentals of web design. It includes mini-projects that emphasize the structure and styling of web pages, responsive layouts, and basic design principles. Through these exercises, I explored HTML5 for semantic markup and CSS3 for creative styling, preparing for more advanced frontend and backend development in the future."